feat: now damages the player
